backend/requirements.txt
# Aegis Backend Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
google-generativeai>=0.3.0
python-multipart>=0.0.9
librosa>=0.10.0
soundfile>=0.12.0

# Database — async PostgreSQL stack
sqlalchemy>=2.0.0
asyncpg>=0.29.0

